Moved utils to utils file
